requirements.txt
Flask>=2.0
Flask-JWT-Extended>=4.7.1
Flask-SQLAlchemy
pdfplumber
python-docx
pytesseract
Pillow
spacy
scikit-learn
requests
PyJWT<2.10
